Reorganized source files
